+++ docs_src/src/guide/changelog.md
@@ -12,6 +12,13 @@ meta:
 
 [[toc]]
 
+## v3.2.0
+
+- Added a way to customize the request's body field used by `JWTController` for the authentication process. (By default, "username" and "password" are used)
+- Added a new validation rule: `unique`.
+- Added `helper.Only()`.
+- `filesystem.File.Save()` now creates directories if needed.
+
 ## v3.1.0
 
 - Added support for validating objects (`map[string]interface{}`).